refactor(gui): split version modals and add reset event
